ai-apps-collection/3D_Video_Edu_Platform/requirements.txt
streamlit>=1.28.0
agno>=0.1.0
crewai>=0.1.0
langchain-openai>=0.1.0
google-generativeai>=0.3.0
bpy>=3.0.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
moviepy>=1.0.3
plotly>=5.15.0
pandas>=2.0.0
asyncio-compat>=0.1.0
aiofiles>=23.0.0
